Fix MySQL timestamp precision. Remove tag cascade deletes. Add entity pretty print functions. Add missing parameters to event triggers.

+++ mysql/build.mysql.sql
@@ -13,7 +13,7 @@ CREATE TABLE org_v (
     org_v_id INTEGER NOT NULL PRIMARY KEY AUTO_INCREMENT, --
     org_id INTEGER NOT NULL, 
     moderation_user_id INTEGER, 
-    a_time FLOAT NOT NULL, 
+    a_time DOUBLE NOT NULL, 
     public BOOLEAN,
     existence BOOLEAN NOT NULL, --
 
@@ -28,7 +28,7 @@ CREATE TABLE orgalias_v (
     orgalias_v_id INTEGER NOT NULL PRIMARY KEY AUTO_INCREMENT, --
     orgalias_id INTEGER NOT NULL, 
     moderation_user_id INTEGER, 
-    a_time FLOAT NOT NULL, 
+    a_time DOUBLE NOT NULL, 
     public BOOLEAN,
     existence BOOLEAN NOT NULL, --
  
@@ -44,7 +44,7 @@ CREATE TABLE event_v (
     event_v_id INTEGER NOT NULL PRIMARY KEY AUTO_INCREMENT, --
     event_id INTEGER NOT NULL, 
     moderation_user_id INTEGER, 
-    a_time FLOAT NOT NULL, 
+    a_time DOUBLE NOT NULL, 
     public BOOLEAN,
     existence BOOLEAN NOT NULL, --
 
@@ -64,17 +64,17 @@ CREATE TABLE address_v (
     address_v_id INTEGER NOT NULL PRIMARY KEY AUTO_INCREMENT, --
     address_id INTEGER NOT NULL, 
     moderation_user_id INTEGER, 
-    a_time FLOAT NOT NULL, 
+    a_time DOUBLE NOT NULL, 
     public BOOLEAN,
     existence BOOLEAN NOT NULL, --
 
     postal LONGTEXT NOT NULL, 
     source LONGTEXT NOT NULL, 
     lookup LONGTEXT, 
-    manual_longitude FLOAT, 
-    manual_latitude FLOAT, 
-    longitude FLOAT, 
-    latitude FLOAT, 
+    manual_longitude DOUBLE, 
+    manual_latitude DOUBLE, 
+    longitude DOUBLE, 
+    latitude DOUBLE, 
 
     KEY moderation_user_id (moderation_user_id),
     CONSTRAINT address_v_c2 FOREIGN KEY (moderation_user_id)
@@ -85,7 +85,7 @@ CREATE TABLE note_v (
     note_v_id INTEGER NOT NULL PRIMARY KEY AUTO_INCREMENT, --
     note_id INTEGER NOT NULL, 
     moderation_user_id INTEGER, 
-    a_time FLOAT NOT NULL, 
+    a_time DOUBLE NOT NULL, 
     public BOOLEAN,
     existence BOOLEAN NOT NULL, --
 
@@ -101,7 +101,7 @@ CREATE TABLE orgtag_v (
     orgtag_v_id INTEGER NOT NULL PRIMARY KEY AUTO_INCREMENT, --
     orgtag_id INTEGER NOT NULL, 
     moderation_user_id INTEGER, 
-    a_time FLOAT NOT NULL, 
+    a_time DOUBLE NOT NULL, 
     public BOOLEAN,
     existence BOOLEAN NOT NULL, --
 
@@ -117,7 +117,7 @@ CREATE TABLE eventtag_v (
     eventtag_v_id INTEGER NOT NULL PRIMARY KEY AUTO_INCREMENT, --
     eventtag_id INTEGER NOT NULL, 
     moderation_user_id INTEGER, 
-    a_time FLOAT NOT NULL, 
+    a_time DOUBLE NOT NULL, 
     public BOOLEAN,
     existence BOOLEAN NOT NULL, --
 
@@ -132,69 +132,69 @@ CREATE TABLE eventtag_v (
 CREATE TABLE org_address_v (
     org_id INTEGER NOT NULL,
     address_id INTEGER NOT NULL, 
-    a_time FLOAT NOT NULL, 
+    a_time DOUBLE NOT NULL, 
     existence BOOLEAN NOT NULL --
 ) DEFAULT CHARSET=utf8;
 
 CREATE TABLE org_note_v (
     org_id INTEGER NOT NULL,
     note_id INTEGER NOT NULL, 
-    a_time FLOAT NOT NULL, 
+    a_time DOUBLE NOT NULL, 
     existence BOOLEAN NOT NULL --
 ) DEFAULT CHARSET=utf8;
 
 CREATE TABLE org_orgtag_v (
     org_id INTEGER NOT NULL,
     orgtag_id INTEGER NOT NULL, 
-    a_time FLOAT NOT NULL, 
+    a_time DOUBLE NOT NULL, 
     existence BOOLEAN NOT NULL --
 ) DEFAULT CHARSET=utf8;
 
 CREATE TABLE orgtag_note_v (
     orgtag_id INTEGER NOT NULL,
     note_id INTEGER NOT NULL, 
-    a_time FLOAT NOT NULL, 
+    a_time DOUBLE NOT NULL, 
     existence BOOLEAN NOT NULL --
 ) DEFAULT CHARSET=utf8;
 
 CREATE TABLE event_address_v (
     event_id INTEGER NOT NULL,
     address_id INTEGER NOT NULL, 
-    a_time FLOAT NOT NULL, 
+    a_time DOUBLE NOT NULL, 
     existence BOOLEAN NOT NULL --
 ) DEFAULT CHARSET=utf8;
 
 CREATE TABLE event_note_v (
     event_id INTEGER NOT NULL,
     note_id INTEGER NOT NULL, 
-    a_time FLOAT NOT NULL, 
+    a_time DOUBLE NOT NULL, 
     existence BOOLEAN NOT NULL --
 ) DEFAULT CHARSET=utf8;
 
 CREATE TABLE event_eventtag_v (
     event_id INTEGER NOT NULL,
     eventtag_id INTEGER NOT NULL, 
-    a_time FLOAT NOT NULL, 
+    a_time DOUBLE NOT NULL, 
     existence BOOLEAN NOT NULL --
 ) DEFAULT CHARSET=utf8;
 
 CREATE TABLE eventtag_note_v (
     eventtag_id INTEGER NOT NULL,
     note_id INTEGER NOT NULL, 
-    a_time FLOAT NOT NULL, 
+    a_time DOUBLE NOT NULL, 
     existence BOOLEAN NOT NULL --
 ) DEFAULT CHARSET=utf8;
 
 CREATE TABLE address_note_v (
     address_id INTEGER NOT NULL,
     note_id INTEGER NOT NULL, 
-    a_time FLOAT NOT NULL, 
+    a_time DOUBLE NOT NULL, 
     existence BOOLEAN NOT NULL --
 ) DEFAULT CHARSET=utf8;
 
 CREATE TABLE org_event_v (
     org_id INTEGER NOT NULL,
     event_id INTEGER NOT NULL, 
-    a_time FLOAT NOT NULL, 
+    a_time DOUBLE NOT NULL, 
     existence BOOLEAN NOT NULL --
 ) DEFAULT CHARSET=utf8;
